requirements.txt
flask
praw
beautifulsoup4
python-dotenv
pytest
requests
